refactor(middleware): move rbac middleware to new file
